+++ src/k8s-extension/azext_k8s_extension/custom.py
@@ -138,7 +138,8 @@ def create_k8s_extension(cmd, client, resource_group_name, cluster_name, name, c
     validate_cc_registration(cmd)
 
     # Create identity, if required
-    if create_identity:
+    # We don't create the identity if we are in DF
+    if create_identity and not __is_dogfood_cluster(cmd):
         extension_instance.identity, extension_instance.location = \
             __create_identity(cmd, resource_group_name, cluster_name, cluster_type, cluster_rp)
 
@@ -292,3 +293,7 @@ def __get_config_settings_from_file(file_path):
         raise Exception("File {} is empty".format(file_path))
 
     return settings
+
+
+def __is_dogfood_cluster(cmd):
+    return cmd.cli_ctx.cloud.endpoints.resource_manager == consts.DF_RM_ENDPOINT